fixed some bugs- we are testing chat, then will move the components to our home page once they work
