+++ react-app/src/App.js
@@ -13,6 +13,7 @@ import CreateForm from "./components/HomePage/createServer";
 import EditServerForm from "./components/HomePage/updateServer";
 import CreateChannel from "./components/HomePage/Channel/createChannel";
 import Splash from "./components/Splash/splash";
+import Chat from "./components/HomePage/Chat/Chat";
 
 function App() {
   const [loaded, setLoaded] = useState(false);
@@ -42,6 +43,9 @@ function App() {
         <Route path="/sign-up" exact={true}>
           <SignUpForm />
         </Route>
+        <ProtectedRoute path="/test-chat" exact={true}>
+          <Chat />
+        </ProtectedRoute>
         <ProtectedRoute path="/users" exact={true}>
           <UsersList />
         </ProtectedRoute>
