+++ react-app/src/components/HomePage/Chat/TestChatPage.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Chat from './Chat'
+
+export default function TestChatPage() {
+  return (
+    <div>
+        <Chat />
+    </div>
+  )
+}
